Split azure module
